+++ src/test/SvLoadLogRecordTest.cpp
@@ -1,5 +1,5 @@
 /**************************************************************************************************
- * Copyright (c) 2022 Calypso Networks Association https://calypsonet.org/                        *
+ * Copyright (c) 2023 Calypso Networks Association https://calypsonet.org/                        *
  *                                                                                                *
  * See the NOTICE file(s) distributed with this work for additional information regarding         *
  * copyright ownership.                                                                           *
@@ -71,7 +71,7 @@ static void setUp()
                                                                     SV_TNUM_STR);
 
     svLoadLogRecordAdapter = std::make_shared<SvLoadLogRecordAdapter>(svGetLoadData,
-                                                                      HEADER.size() / 2);
+                                                                      static_cast<int>(HEADER.size() / 2));
 }
 
 static void tearDown()
